+++ FE/src/App.tsx
@@ -1,11 +1,11 @@
 import { ThemeProvider } from 'styled-components';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import LoginPage from '@/login/LoginPage';
-import IssueListPage from '@/issueList/IssueListPage';
+import LoginPage from '@/pages/LoginPage';
+import IssueListPage from '@/pages/IssueListPage';
+import { lazy, Suspense } from 'react';
 import GlobalStyles from './GlobalStyles';
 import theme from './theme';
-import Layout from './common/Layout';
-import { lazy, Suspense } from 'react';
+import Layout from './common/components/Layout';
 
 const pages = ['addIssue', 'issueDetail', 'labelList', 'milestoneList'].sort();
 
@@ -17,9 +17,13 @@ const getFileName = (page: string) => {
 
 const getPage = (page: string) => {
   const fileName = getFileName(page);
-  const Page = lazy(() => import('./pages/' + fileName));
+  const Page = lazy(() => import(`./pages/${fileName}`));
 
-  return <Suspense fallback={<div>Loading...</div>}>{<Page />}</Suspense>;
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <Page />
+    </Suspense>
+  );
 };
 
 export default function App() {
