+++ FE/src/common/components/dropDown/DropDownPanel.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { GREYSCALE } from '@/constants';
+import { GREYSCALE } from '@/common/constants';
 import DropDownCheckBoxItem from './DropDownCheckBoxItem';
 
 type DropDownPanelProps = {
